requirements.txt
opencv2
easyocr
pandas